# packages/schema/openapi/valuation.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "not_found",
              "conflict",
              "invalid_request"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "const": "N/A",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "AssumptionContent": {
        "additionalProperties": false,
        "properties": {
          "anchor_method": {
            "const": "explicit_carry_forward_annual_run_rate",
            "title": "Anchor Method",
            "type": "string"
          },
          "anchor_source_index": {
            "minimum": 0.0,
            "title": "Anchor Source Index",
            "type": "integer"
          },
          "author_id": {
            "maxLength": 250,
            "minLength": 1,
            "title": "Author Id",
            "type": "string"
          },
          "authored_at": {
            "format": "date-time",
            "title": "Authored At",
            "type": "string"
          },
          "calendar": {
            "const": "calendar_year_equivalent",
            "title": "Calendar",
            "type": "string"
          },
          "forecast_dates": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Forecast Dates",
            "type": "array"
          },
          "judgments": {
            "items": {
              "$ref": "#/components/schemas/Judgment"
            },
            "title": "Judgments",
            "type": "array"
          },
          "leap_day": {
            "const": "february_28",
            "title": "Leap Day",
            "type": "string"
          },
          "retrospective": {
            "title": "Retrospective",
            "type": "boolean"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/Scenario"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Scenarios",
            "type": "array"
          },
          "sensitivities": {
            "items": {
              "$ref": "#/components/schemas/SensitivityGrid"
            },
            "maxItems": 4,
            "title": "Sensitivities",
            "type": "array"
          },
          "valuation_at": {
            "format": "date-time",
            "title": "Valuation At",
            "type": "string"
          }
        },
        "required": [
          "authored_at",
          "author_id",
          "valuation_at",
          "retrospective",
          "calendar",
          "leap_day",
          "forecast_dates",
          "anchor_method",
          "anchor_source_index",
          "scenarios",
          "sensitivities",
          "judgments"
        ],
        "title": "AssumptionContent",
        "type": "object"
      },
      "AssumptionCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/AssumptionContent"
          },
          "idempotency_key": {
            "maxLength": 250,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "parent_id",
          "idempotency_key",
          "content"
        ],
        "title": "AssumptionCreate",
        "type": "object"
      },
      "AssumptionSaved": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/AssumptionContent"
          },
          "content_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "id",
          "parent_id",
          "content_hash",
          "content"
        ],
        "title": "AssumptionSaved",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Interval": {
        "additionalProperties": false,
        "properties": {
          "lower": {
            "title": "Lower",
            "type": "string"
          },
          "upper": {
            "title": "Upper",
            "type": "string"
          }
        },
        "required": [
          "lower",
          "upper"
        ],
        "title": "Interval",
        "type": "object"
      },
...
      "LatestResponse": {
        "additionalProperties": false,
        "properties": {
          "latest_request_id": {
            "format": "uuid",
            "title": "Latest Request Id",
            "type": "string"
          },
          "latest_request_state": {
            "title": "Latest Request State",
            "type": "string"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run",
          "latest_request_id",
          "latest_request_state"
        ],
        "title": "LatestResponse",
        "type": "object"
      },
...
      "RequestCreate": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 250,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "max_attempts": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "parent_request_id": {
            "format": "uuid",
            "title": "Parent Request Id",
            "type": "string"
          },
          "review_id": {
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "parent_request_id",
          "idempotency_key",
          "max_attempts"
        ],
        "title": "RequestCreate",
        "type": "object"
      },
      "RequestResponse": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "worker_failure",
                  "cancelled",
                  "lease_expired",
                  "waiting_for_input",
                  "retry_scheduled",
                  "stage_blocked",
                  "stage_unsupported",
                  "stage_failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "execution_id": {
            "format": "uuid",
            "title": "Execution Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "stage_key": {
            "anyOf": [
              {
                "const": "valuation",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Key"
          },
          "stage_state": {
            "anyOf": [
              {
                "enum": [
                  "running",
                  "completed",
                  "blocked",
                  "unsupported",
                  "failed",
                  "cancelled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage State"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "waiting_for_input",
              "retry_scheduled",
              "completed",
              "completed_with_gaps",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "execution_id",
          "state",
          "stage_key",
          "stage_state",
          "error_code",
          "run_id"
        ],
        "title": "RequestResponse",
        "type": "object"
      },
      "RunResponse": {
        "additionalProperties": false,
        "properties": {
          "compatibility_key": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Compatibility Key",
            "type": "string"
          },
          "execution_id": {
            "format": "uuid",
            "title": "Execution Id",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "input_snapshot_id": {
            "format": "uuid",
            "title": "Input Snapshot Id",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "completed",
              "completed_with_gaps"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "payload_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Payload Hash",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ValuationPayload"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "parent_run_id",
          "request_id",
          "execution_id",
          "input_snapshot_id",
          "generated_at",
          "compatibility_key",
          "payload_hash",
          "outcome",
          "result"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValuationPayload": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/SafeAssumptions"
          },
          "bridge": {
            "$ref": "#/components/schemas/BridgeResult"
          },
          "bridge_captured_before": {
            "format": "date-time",
            "title": "Bridge Captured Before",
            "type": "string"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "conclusions": {
            "items": {
              "type": "string"
            },
            "title": "Conclusions",
            "type": "array"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "dependency_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Dependency Hash",
            "type": "string"
          },
          "evidence_mode": {
            "const": "synthetic",
            "title": "Evidence Mode",
            "type": "string"
          },
          "fixture_label": {
            "const": "Synthetic integration fixture — not company data",
            "title": "Fixture Label",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/ModelDefinition"
          },
          "numerical_uncertainty": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "items": {
                    "$ref": "#/components/schemas/Interval"
                  },
                  "type": "array"
                }
              ],
              "type": "array"
            },
            "title": "Numerical Uncertainty",
            "type": "array"
          },
          "policy": {
            "$ref": "#/components/schemas/ValuationPolicy"
          },
          "price": {
            "$ref": "#/components/schemas/PriceEvidence"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "scenario_dispersion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Interval"
              },
              {
                "type": "null"
              }
            ]
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioResult"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "sensitivities": {
            "items": {
              "$ref": "#/components/schemas/SensitivityResult"
            },
            "title": "Sensitivities",
            "type": "array"
          },
          "shares": {
            "$ref": "#/components/schemas/SharePool"
          },
          "source_eligibility_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Source Eligibility Reasons",
            "type": "array"
          },
          "source_input_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Source Input Hash",
            "type": "string"
          },
          "source_measurement_uncertainty": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "maxLength": 250,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "anyOf": [
                    {
                      "format": "decimal_dcf_input",
                      "pattern": "^-?(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?(?:[eE][+-]?[0-9]+)?$",
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              ],
              "type": "array"
            },
            "title": "Source Measurement Uncertainty",
            "type": "array"
          },
          "source_payload_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Source Payload Hash",
            "type": "string"
          },
          "source_selector": {
            "$ref": "#/components/schemas/SelectionKey"
          },
          "source_snapshot_id": {
            "format": "uuid",
            "title": "Source Snapshot Id",
            "type": "string"
          },
          "valuation_at": {
            "format": "date-time",
            "title": "Valuation At",
            "type": "string"
          },
          "version": {
            "const": "s7-result-v1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "evidence_mode",
          "fixture_label",
          "dependency_hash",
          "model",
          "policy",
          "assumptions",
          "source_snapshot_id",
          "source_payload_hash",
          "source_input_hash",
          "source_selector",
          "price",
          "shares",
          "claims",
          "bridge_captured_before",
          "valuation_at",
          "currency",
          "bridge",
          "scenarios",
          "scenario_dispersion",
          "numerical_uncertainty",
          "source_eligibility_reasons",
          "source_measurement_uncertainty",
          "sensitivities",
          "conclusions",
          "reasons"
        ],
        "title": "ValuationPayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "equityEval Synthetic Valuation",
    "version": "s7a-v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/assumption-sets": {
      "post": {
        "operationId": "new_assumptions_assumption_sets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssumptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssumptionSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Assumptions"
      }
    },
    "/assumption-sets/{assumption_id}": {
      "get": {
        "operationId": "assumptions_assumption_sets__assumption_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assumption_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assumption Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssumptionSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assumptions"
      }
    },
    "/companies/{issuer_id}/valuation": {
      "get": {
        "operationId": "latest_companies__issuer_id__valuation_get",
        "parameters": [
          {
            "in": "path",
            "name": "issuer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issuer Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "security_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Security Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quote_identifier_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Quote Identifier Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "compatibility_key",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{64}$",
              "title": "Compatibility Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest"
      }
    },
    "/model-runs/{run_id}": {
      "get": {
        "operationId": "saved_run_model_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Saved Run"
      }
    },
    "/valuation-requests": {
      "post": {
        "operationId": "enqueue_valuation_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue"
      }
    },
    "/valuation-requests/{request_id}": {
      "get": {
        "operationId": "request_state_valuation_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request State"
      }
    }
  }
}